refactor EpisodeInfo - do not index given file anymore, accept videonode or src_file - updated clip retrival API to match src_file (src, src_cut, init, init_cut) - add option to get untrimmed ncop/nced
