+++ svsfunc/__init__.py
@@ -1,6 +1,6 @@
 from .bdmv import *  # noqa: F401, F403
 from .filterchain import *  # noqa: F401, F403
-from .indexer import *  # noqa: F401, F403
+from .episode_info import *  # noqa: F401, F403
 from .parse import *  # noqa: F401, F403
 from .utils import *  # noqa: F401, F403
